增加直接导出markdown的功能，其中slug输出后自定义

+++ app.R
@@ -5,7 +5,7 @@ library(stringr)
 author = "shitao5"
 
 add_yaml <- function(file, title, author, date, format) {
-  # 定义基础的 YAML 内容
+  # 定义为 quarto 使用的基础 YAML 内容
   basic_yaml <- paste0(
     "---\n",
     "title: \"", title, "\"\n",
@@ -43,8 +43,17 @@ add_yaml <- function(file, title, author, date, format) {
       "    number-sections: true\n",
       "---\n\n"
     )
+  } else if (tolower(format) == "markdown") {
+    yaml_content <- paste0(
+      "---\n",
+      "title: \"", title, "\"\n",
+      "author: \"", author, "\"\n",
+      "date: \"", date, "\"\n",
+      "slug: \"", "\"\n",
+      "---\n\n"
+    )
   } else {
-    stop("Unsupported format. Please choose 'html' or 'word'.")
+    stop("Unsupported format. Please choose 'html', 'docx' or 'markdown'.")
   }
   
   c(yaml_content, file)
@@ -63,7 +72,7 @@ server <- function(input, output, session) {
   # 监听按钮点击事件，进行处理和渲染
   output$download <- downloadHandler(
     filename = function() {
-      paste0(tools::file_path_sans_ext(input$file$name), ifelse(input$format == "html", ".html", ".docx"))
+      paste0(tools::file_path_sans_ext(input$file$name), ifelse(input$format == "html", ".html", ifelse(input$format == "markdown", ".md", ".docx")))
     },
     content = function(file) {
       req(file_data())
@@ -78,28 +87,31 @@ server <- function(input, output, session) {
                       date = input$date,
                       format = input$format)
       
-      # 写入到临时文件
-      temp_qmd <- tempfile(fileext = ".qmd")
-      temp_dir = dirname(temp_qmd)
-      writeLines(res, temp_qmd)
-      file.copy("pangu.min.js", temp_dir)
-      
-      # 设置输出格式和文件名
-      output_format <- ifelse(input$format == "html", "html", "docx")
-      output_file_name <- paste0(temp_dir, "/",
-                                 tools::file_path_sans_ext(basename(temp_qmd)), ".", output_format)
-      
-      # 调用系统命令执行 Quarto 渲染
-      cmd <- paste("quarto render", shQuote(temp_qmd))
-      system(cmd)
-      
-      # 将渲染完成的文件复制到 Shiny 输出文件位置
-      file.copy(output_file_name, file, overwrite = TRUE)
-      
-      # 删除临时文件
-      unlink(c(temp_qmd, output_file_name))
+      if (input$format == "markdown") {
+        writeLines(res, file)
+      } else {
+        # 写入到临时文件
+        temp_qmd <- tempfile(fileext = ".qmd")
+        temp_dir = dirname(temp_qmd)
+        writeLines(res, temp_qmd)
+        file.copy("pangu.min.js", temp_dir)
+        
+        # 设置输出格式和文件名
+        output_format <- ifelse(input$format == "html", "html", "docx")
+        output_file_name <- paste0(temp_dir, "/",
+                                   tools::file_path_sans_ext(basename(temp_qmd)), ".", output_format)
+        
+        # 调用系统命令执行 Quarto 渲染
+        cmd <- paste("quarto render", shQuote(temp_qmd))
+        system(cmd)
+        
+        # 将渲染完成的文件复制到 Shiny 输出文件位置
+        file.copy(output_file_name, file, overwrite = TRUE)
+        
+        # 删除临时文件
+        unlink(c(temp_qmd, output_file_name))
+      }
     }
-    
   )
 }
 
@@ -112,7 +124,7 @@ ui <- fluidPage(
       textInput("title", "标题", value = "", placeholder = "默认为文件名"),
       textInput("author", "作者", value = author),
       dateInput("date", "日期", value = Sys.Date()),
-      selectInput("format", "输出格式", choices = c("html", "docx")),
+      selectInput("format", "输出格式", choices = c("html", "docx", "markdown")),
       downloadButton("download", "渲染并下载")
     ),
     mainPanel(
